ChartView screen finished

+++ app/src/main/java/com/example/stocks/models/remote/StockChart.kt
@@ -8,5 +8,5 @@ data class StockChart(
     @SerializedName("low") var low: Float? = null,
     @SerializedName("high") var high: Float? = null,
     @SerializedName("close") var close: Float? = null,
-    @SerializedName("volume") var volume: Int?    = null
+    @SerializedName("volume") var volume: Float?    = null
 )
